[deckhouse-cli] Drop the direct-registry plugin service, keep the contract decoder
- remove PluginService and its OCI methods (contract-from-annotation, image extract, tag/catalog listing); the registry-packages-proxy is the only plugin source
- keep the contract decoder and DTO<->domain converters in contract.go, reused by the rpp source, validators and install
